Rename 'dellemc' to 'dell_emc' as in opensource drivers
